fix: fix errors in input
