+++ src/components/Header/Header.module.css
@@ -25,3 +25,19 @@
 .title_wrapper {
   text-decoration: none;
 }
+
+.form {
+  display: flex;
+  gap: 10px;
+  padding-inline: 5px;
+}
+
+.input {
+  padding: 5px 10px;
+  font-size: 14px;
+  width: 300px;
+  border-radius: 20px;
+  border: 1px solid #000000;
+  background-color: rgba(255, 240, 0, 0.1);
+  color: #000000;
+}
