+++ src/components/Header/Header.tsx
@@ -1,27 +1,29 @@
 import { useRef } from 'react';
 import { Link, useSearchParams } from 'react-router-dom';
 import { DEFAULT_PAGE, SEARCH_PARAM_PAGE } from '../../models/constants';
-import SearchInput from '../SearchInput/SearchInput';
 import styles from './Header.module.css';
 import { setLocalStorage } from '../../LocalStorage/setLocalStorage';
 import { pageCurrentUpdate } from '../../Store/Reducers/PageCurrentSlice';
+import { useAppDispatch, useAppSelector } from '../../hooks/reduxHoooks';
+import {
+  selectSearch,
+  setRootSearch,
+} from '../../Store/Reducers/SearchReduser';
+import SearchButton from '../SearchInput/SearchButton/SearchButton';
 
 const Header = () => {
   const textInput = useRef<HTMLInputElement>(null);
   const [searchParams, setSearchParams] = useSearchParams();
-
+  const dispatch = useAppDispatch();
+  const searchTerm = useAppSelector(selectSearch);
   const handleSearchClick = () => {
     const value = textInput.current?.value.trim() || '';
     setLocalStorage(value);
-
-    pageCurrentUpdate(DEFAULT_PAGE);
+    dispatch(setRootSearch(value));
+    dispatch(pageCurrentUpdate(DEFAULT_PAGE));
 
     searchParams.delete(SEARCH_PARAM_PAGE);
     setSearchParams(searchParams);
-  };
-
-  const handleTitleClick = () => {
-    pageCurrentUpdate(DEFAULT_PAGE);
   };
 
   const handleButtonClick = () => {
@@ -33,12 +35,19 @@ const Header = () => {
 
   return (
     <header className={styles.header}>
-      <Link to="/" className={styles.title_wrapper} onClick={handleTitleClick}>
+      <Link to="/" className={styles.title_wrapper}>
         <h1 className={styles.title} onClick={handleButtonClick}>
           Star Wars
         </h1>
       </Link>
-      <SearchInput />
+      <div className={styles.form}>
+        <input
+          ref={textInput}
+          defaultValue={searchTerm}
+          className={styles.input}
+        />
+        <SearchButton onClick={handleSearchClick} />
+      </div>
     </header>
   );
 };
